Create RPI_Receiver.py
Save the Python script and run it on your Raspberry Pi. It will continuously listen for data from the Arduino over I2C and print the received messages.

In this example, the Arduino is acting as the I2C master and is sending messages to the Raspberry Pi, which acts as the I2C slave. The Raspberry Pi reads the data sent by the Arduino using the bus.read_i2c_block_data() function.

Make sure to connect the SDA (data) and SCL (clock) pins of the Arduino to the corresponding SDA and SCL pins on the Raspberry Pi. Also, ensure that both devices share a common ground.

Again, please verify voltage compatibility, connections, and refer to the documentation of both devices for accurate setup details.