Create Arduino_Receiver.ino
In this example, the Raspberry Pi is acting as the I2C master, and the Arduino is acting as the I2C slave. The Raspberry Pi sends messages using the bus.write_i2c_block_data() function, and the Arduino reads the messages using the Wire.read() function.

Make sure to connect the SDA (data) and SCL (clock) pins of the Raspberry Pi to the corresponding SDA and SCL pins on the Arduino. Also, ensure that both devices share a common ground.